+++ app/src/main/java/com/mapbox/maps/testapp/examples/markersandcallouts/viewannotation/DynamicViewAnnotationActivity.kt
@@ -177,6 +177,7 @@ class DynamicViewAnnotationActivity : AppCompatActivity() {
     }
   }
 
+  @OptIn(com.mapbox.maps.MapboxExperimental::class)
   private fun MapboxMap.initStyleWithLayers(onLoaded: (Style) -> Unit) {
     refreshRoutes()
 
@@ -189,6 +190,9 @@ class DynamicViewAnnotationActivity : AppCompatActivity() {
         // layer for alternative route
         +routeLayerAlt.apply {
           lineColor(Color.parseColor("#FF999999"))
+          // make map elements around (30m) route line visible through obstructing 3D buildings and other aboveground features
+          lineCutoutWidth(30.0)
+          lineCutoutOpacity(0.2)
           lineWidth(12.0)
           lineBorderWidth(2.0)
           lineBorderColor(Color.parseColor("#FF333333"))
@@ -196,6 +200,9 @@ class DynamicViewAnnotationActivity : AppCompatActivity() {
         // layer for main route
         +routeLayerMain.apply {
           lineColor(Color.parseColor("#FF57A9FB"))
+          // make map elements around (30m) route line visible through obstructing 3D buildings and other aboveground features
+          lineCutoutWidth(30.0)
+          lineCutoutOpacity(0.2)
           lineWidth(12.0)
           lineCap(LineCap.ROUND)
           lineBorderWidth(2.0)